+++ packages/contracts-js/scripts/extract-abis.mjs
@@ -9,15 +9,20 @@ import {
 import { basename, join, resolve } from "path";
 
 const main = () => {
-  const libFolder = resolve(
+  const sourceFolders = [];
+  sourceFolders.push(resolve(
     "node_modules",
     "@thirdweb-dev/contracts/abi",
-  );
+  ));
+  sourceFolders.push(resolve(
+    "external_abis",
+  ));
   const destinationFolder = "./dist/abis";
   if (existsSync(destinationFolder)) {
     rmSync(destinationFolder, { recursive: true });
   }
   mkdirSync(destinationFolder);
+  for(const libFolder of sourceFolders) {
   const abiFiles = readdirSync(libFolder);
   for (const abiFile of abiFiles) {
     const contractJsonFile = readFileSync(join(libFolder, abiFile), "utf-8");
@@ -28,6 +33,7 @@ const main = () => {
       JSON.stringify(abi, null, 2),
     );
   }
+ }
 };
 
 main();