+++ packages/sdk/src/evm/common/any-evm-utils/getInitBytecodeWithSalt.ts
@@ -13,8 +13,9 @@ import { getSaltHash } from "./getSaltHash";
 export function getInitBytecodeWithSalt(
   bytecode: string,
   encodedArgs: BytesLike,
+  salt?: string,
 ): string {
-  const saltHash = getSaltHash(bytecode);
+  const saltHash = salt ? ethers.utils.id(salt) : getSaltHash(bytecode);
 
   const initBytecodeWithSalt = ethers.utils.solidityPack(
     ["bytes32", "bytes", "bytes"],