+++ packages/sdk/src/evm/constants/contract.ts
@@ -6,9 +6,10 @@ import { utils } from "ethers";
 /**
  * @internal
  */
-export const InterfaceId_IERC721 = utils.arrayify("0x80ac58cd");
+export const InterfaceId_IERC721 = /* @__PURE__ */ utils.arrayify("0x80ac58cd");
 
 /**
  * @internal
  */
-export const InterfaceId_IERC1155 = utils.arrayify("0xd9b67a26");
+export const InterfaceId_IERC1155 =
+  /* @__PURE__ */ utils.arrayify("0xd9b67a26");