+++ packages/sdk/src/evm/schema/contracts/common/airdrop.ts
@@ -5,24 +5,26 @@ import { z } from "zod";
 /**
  * @internal
  */
-export const AirdropAddressInput = z.object({
-  address: AddressOrEnsSchema,
-  quantity: AmountSchema.default(1),
-});
+export const AirdropAddressInput = /* @__PURE__ */ (() =>
+  z.object({
+    address: AddressOrEnsSchema,
+    quantity: AmountSchema.default(1),
+  }))();
 
 /**
  * @internal
  */
-export const AirdropInputSchema = z.union([
-  z.array(z.string()).transform(
-    async (strings) =>
-      await Promise.all(
-        strings.map((address) =>
-          AirdropAddressInput.parseAsync({
-            address,
-          }),
+export const AirdropInputSchema = /* @__PURE__ */ (() =>
+  z.union([
+    z.array(z.string()).transform(
+      async (strings) =>
+        await Promise.all(
+          strings.map((address) =>
+            AirdropAddressInput.parseAsync({
+              address,
+            }),
+          ),
         ),
-      ),
-  ),
-  z.array(AirdropAddressInput),
-]);
+    ),
+    z.array(AirdropAddressInput),
+  ]))();