+++ packages/sdk/src/evm/schema/shared/CallOverrideSchema.ts
@@ -3,16 +3,18 @@ import { z } from "zod";
 import { AddressOrEnsSchema } from "./AddressOrEnsSchema";
 import { BigNumberishSchema } from "./BigNumberSchema";
 
-export const CallOverrideSchema: z.ZodType<CallOverrides> = z
-  .object({
-    gasLimit: BigNumberishSchema.optional(),
-    gasPrice: BigNumberishSchema.optional(),
-    maxFeePerGas: BigNumberishSchema.optional(),
-    maxPriorityFeePerGas: BigNumberishSchema.optional(),
-    nonce: BigNumberishSchema.optional(),
-    value: BigNumberishSchema.optional(),
-    blockTag: z.union([z.string(), z.number()]).optional(),
-    from: AddressOrEnsSchema.optional(),
-    type: z.number().optional(),
-  })
-  .strict();
+export const CallOverrideSchema: z.ZodType<CallOverrides> =
+  /* @__PURE__ */ (() =>
+    z
+      .object({
+        gasLimit: BigNumberishSchema.optional(),
+        gasPrice: BigNumberishSchema.optional(),
+        maxFeePerGas: BigNumberishSchema.optional(),
+        maxPriorityFeePerGas: BigNumberishSchema.optional(),
+        nonce: BigNumberishSchema.optional(),
+        value: BigNumberishSchema.optional(),
+        blockTag: z.union([z.string(), z.number()]).optional(),
+        from: AddressOrEnsSchema.optional(),
+        type: z.number().optional(),
+      })
+      .strict())();